change shot time type from long to double

+++ capstone/src/main/java/com/google/sps/data/Shot.java
@@ -16,10 +16,10 @@ package com.google.sps.data;
 
 /** A detected shot, with start and end time offsets (seconds)**/
 public class Shot {
-  private final long start_time;
-  private final long end_time;
+  private final double start_time;
+  private final double end_time;
 
-  public Shot(long start, long end) {
+  public Shot(double start, double end) {
     this.start_time = start;
     this.end_time = end;
   }
